+++ fsm/__init__.py
@@ -1,6 +1,6 @@
 """Initializing package."""
 
 from .exceptions import InvalidTransition  # NOQA
-from .fsm import FiniteStateMachineMixin  # NOQA
+from .fsm import FiniteStateMachineMixin, BaseFiniteStateMachineMixin  # NOQA
 
 __version__ = "1.0.1"